V0.5.0

+++ lib/theme/values.dart
@@ -1,5 +1,5 @@
 const int heightNavBar = 64;
-const int borderRadiusNavBar = 20;
+const double borderRadiusNavBar = 20;
 const int widthTapItemNavBar = 58;
 const int sizeIconNavBar = 24;
 const int horizontalMarginNavBar = 26;
